moved the steps to the correct file
